requirements.txt
python-nmap
## Others can install everything with pip install -r requirements.txt ##
